agents/requirements.txt
google-adk>=0.1.0
pydantic>=2.0.0
google-genai>=0.2.0
python-dotenv
requests>=2.31.0>=1.0.0
